Move generic table update method to DB class.
